requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-api-response==0.1.0
django-redis>=5.4.0
django-environ==0.9.0
django-ckeditor==6.3.1

pillow==11.0.0

uvicorn>=0.23.0

channels==4.1.0
channels-redis==4.2.0

whitenoise==6.8.2

psycopg2==2.9.4

celery==5.4.0
django-celery-results==2.5.1
django-celery-beat==2.7.0

Faker==33.0.0